+++ src/OctoshiftCLI.Tests/ado2gh/Commands/RewirePipelineCommandTests.cs
@@ -25,7 +25,7 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
         {
             Assert.NotNull(_command);
             Assert.Equal("rewire-pipeline", _command.Name);
-            Assert.Equal(8, _command.Options.Count);
+            Assert.Equal(9, _command.Options.Count);
 
             TestHelpers.VerifyCommandOption(_command.Options, "ado-org", true);
             TestHelpers.VerifyCommandOption(_command.Options, "ado-team-project", true);
@@ -34,6 +34,7 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
             TestHelpers.VerifyCommandOption(_command.Options, "github-repo", true);
             TestHelpers.VerifyCommandOption(_command.Options, "service-connection-id", true);
             TestHelpers.VerifyCommandOption(_command.Options, "ado-pat", false);
+            TestHelpers.VerifyCommandOption(_command.Options, "default-branch", false);
             TestHelpers.VerifyCommandOption(_command.Options, "verbose", false);
         }
 
@@ -68,9 +69,63 @@ namespace OctoshiftCLI.Tests.AdoToGithub.Commands
 
             _mockAdoApiFactory.Setup(m => m.Create(adoPat)).Returns(_mockAdoApi.Object);
 
-            await _command.Invoke("adoOrg", "adoTeamProject", "adoPipeline", "githubOrg", "githubRepo", "serviceConnectionId", adoPat);
+            await _command.Invoke("adoOrg", "adoTeamProject", "adoPipeline", "githubOrg", "githubRepo", "serviceConnectionId", adoPat: adoPat);
 
             _mockAdoApiFactory.Verify(m => m.Create(adoPat));
+        }
+
+        [Fact]
+        public async Task Command_Uses_Existing_Default_Branch_When_Option_Not_Present()
+        {
+            var adoOrg = "FooOrg";
+            var adoTeamProject = "BlahTeamProject";
+            var adoPipeline = "foo-pipeline";
+            var githubOrg = "foo-gh-org";
+            var githubRepo = "foo-repo";
+            var serviceConnectionId = Guid.NewGuid().ToString();
+            var pipelineId = 1234;
+            var defaultBranch = "default-branch";
+            var clean = "true";
+            var checkoutSubmodules = "null";
+
+            const string adoPat = "ado-pat";
+
+            _mockAdoApi.Setup(x => x.GetPipelineId(adoOrg, adoTeamProject, adoPipeline).Result).Returns(pipelineId);
+            _mockAdoApi.Setup(x => x.GetPipeline(adoOrg, adoTeamProject, pipelineId).Result).Returns((defaultBranch, clean, checkoutSubmodules));
+            _mockAdoApiFactory.Setup(m => m.Create(adoPat)).Returns(_mockAdoApi.Object);
+
+
+            await _command.Invoke(adoOrg, adoTeamProject, adoPipeline, githubOrg, githubRepo, serviceConnectionId, defaultBranch: null, adoPat: adoPat);
+
+            _mockAdoApi.Verify(x => x.ChangePipelineRepo(adoOrg, adoTeamProject, pipelineId, defaultBranch, clean, checkoutSubmodules, githubOrg, githubRepo, serviceConnectionId));
+        }
+
+        /// <summary>
+        /// Ensures that when the default branch option is specified the old pipeline default branch value is not used.
+        /// </summary>
+        [Fact]
+        public async Task Command_Uses_Default_Branch_Option_Not_Existing_Default_Branch()
+        {
+            var adoOrg = "FooOrg";
+            var adoTeamProject = "BlahTeamProject";
+            var adoPipeline = "foo-pipeline";
+            var githubOrg = "foo-gh-org";
+            var githubRepo = "foo-repo";
+            var serviceConnectionId = Guid.NewGuid().ToString();
+            var pipelineId = 1234;
+            var defaultBranch = "default-branch";
+            var clean = "true";
+            var checkoutSubmodules = "null";
+
+            const string adoPat = "ado-pat";
+
+            _mockAdoApi.Setup(x => x.GetPipelineId(adoOrg, adoTeamProject, adoPipeline).Result).Returns(pipelineId);
+            _mockAdoApi.Setup(x => x.GetPipeline(adoOrg, adoTeamProject, pipelineId).Result).Returns(("old-default-branch", clean, checkoutSubmodules));
+            _mockAdoApiFactory.Setup(m => m.Create(adoPat)).Returns(_mockAdoApi.Object);
+
+            await _command.Invoke(adoOrg, adoTeamProject, adoPipeline, githubOrg, githubRepo, serviceConnectionId, adoPat: adoPat, defaultBranch: defaultBranch);
+
+            _mockAdoApi.Verify(x => x.ChangePipelineRepo(adoOrg, adoTeamProject, pipelineId, defaultBranch, clean, checkoutSubmodules, githubOrg, githubRepo, serviceConnectionId));
         }
     }
 }